split the vpc endpoint stuff from the rest of the crd code
this will allow us to more easily use the materialize crd from cloud
