+++ src/workspace-hack/Cargo.toml
@@ -57,7 +57,9 @@ futures-util = { version = "0.3.30", features = ["channel", "io", "sink"] }
 getrandom = { version = "0.2.10", default-features = false, features = ["std"] }
 globset = { version = "0.4.14", features = ["serde1"] }
 hashbrown = { version = "0.14.5", features = ["raw"] }
-hyper = { version = "0.14.27", features = ["client", "http1", "http2", "stream", "tcp"] }
+hyper-582f2526e08bb6a0 = { package = "hyper", version = "0.14.27", features = ["client", "http1", "http2", "stream", "tcp"] }
+hyper-dff4ba8e3ae991db = { package = "hyper", version = "1.4.1", features = ["client", "http1", "http2", "server"] }
+hyper-util = { version = "0.1.6", features = ["client-legacy", "server-auto", "service"] }
 indexmap = { version = "1.9.1", default-features = false, features = ["std"] }
 insta = { version = "1.33.0", features = ["json"] }
 k8s-openapi = { version = "0.22.0", default-features = false, features = ["schemars", "v1_29"] }
@@ -82,6 +84,7 @@ num-bigint = { version = "0.4.3" }
 num-integer = { version = "0.1.46", features = ["i128"] }
 num-traits = { version = "0.2.15", features = ["i128", "libm"] }
 openssl = { version = "0.10.66", features = ["vendored"] }
+openssl-sys = { version = "0.9.103", default-features = false, features = ["vendored"] }
 ordered-float = { version = "4.2.0", features = ["serde"] }
 parking_lot = { version = "0.12.1", features = ["send_guard"] }
 parquet = { version = "51.0.0", default-features = false, features = ["arrow", "brotli", "flate2", "lz4", "snap", "zstd"] }
@@ -184,7 +187,9 @@ futures-util = { version = "0.3.30", features = ["channel", "io", "sink"] }
 getrandom = { version = "0.2.10", default-features = false, features = ["std"] }
 globset = { version = "0.4.14", features = ["serde1"] }
 hashbrown = { version = "0.14.5", features = ["raw"] }
-hyper = { version = "0.14.27", features = ["client", "http1", "http2", "stream", "tcp"] }
+hyper-582f2526e08bb6a0 = { package = "hyper", version = "0.14.27", features = ["client", "http1", "http2", "stream", "tcp"] }
+hyper-dff4ba8e3ae991db = { package = "hyper", version = "1.4.1", features = ["client", "http1", "http2", "server"] }
+hyper-util = { version = "0.1.6", features = ["client-legacy", "server-auto", "service"] }
 indexmap = { version = "1.9.1", default-features = false, features = ["std"] }
 insta = { version = "1.33.0", features = ["json"] }
 k8s-openapi = { version = "0.22.0", default-features = false, features = ["schemars", "v1_29"] }
@@ -209,6 +214,7 @@ num-bigint = { version = "0.4.3" }
 num-integer = { version = "0.1.46", features = ["i128"] }
 num-traits = { version = "0.2.15", features = ["i128", "libm"] }
 openssl = { version = "0.10.66", features = ["vendored"] }
+openssl-sys = { version = "0.9.103", default-features = false, features = ["vendored"] }
 ordered-float = { version = "4.2.0", features = ["serde"] }
 parking_lot = { version = "0.12.1", features = ["send_guard"] }
 parquet = { version = "51.0.0", default-features = false, features = ["arrow", "brotli", "flate2", "lz4", "snap", "zstd"] }
@@ -271,25 +277,23 @@ zstd-sys = { version = "2.0.9", features = ["std"] }
 [target.x86_64-unknown-linux-gnu.dependencies]
 bitflags = { version = "2.4.1", default-features = false, features = ["std"] }
 camino = { version = "1.1.7", default-features = false, features = ["serde1"] }
-hyper = { version = "0.14.27", default-features = false, features = ["runtime"] }
+hyper-582f2526e08bb6a0 = { package = "hyper", version = "0.14.27", default-features = false, features = ["runtime"] }
 native-tls = { version = "0.2.11", default-features = false, features = ["vendored"] }
-openssl-sys = { version = "0.9.103", default-features = false, features = ["vendored"] }
 pathdiff = { version = "0.2.1", default-features = false, features = ["camino"] }
 ring = { version = "0.17.7", features = ["std"] }
 
 [target.x86_64-unknown-linux-gnu.build-dependencies]
 bitflags = { version = "2.4.1", default-features = false, features = ["std"] }
 camino = { version = "1.1.7", default-features = false, features = ["serde1"] }
-hyper = { version = "0.14.27", default-features = false, features = ["runtime"] }
+hyper-582f2526e08bb6a0 = { package = "hyper", version = "0.14.27", default-features = false, features = ["runtime"] }
 native-tls = { version = "0.2.11", default-features = false, features = ["vendored"] }
-openssl-sys = { version = "0.9.103", default-features = false, features = ["vendored"] }
 pathdiff = { version = "0.2.1", default-features = false, features = ["camino"] }
 ring = { version = "0.17.7", features = ["std"] }
 
 [target.x86_64-apple-darwin.dependencies]
 bitflags = { version = "2.4.1", default-features = false, features = ["std"] }
 camino = { version = "1.1.7", default-features = false, features = ["serde1"] }
-hyper = { version = "0.14.27", default-features = false, features = ["runtime"] }
+hyper-582f2526e08bb6a0 = { package = "hyper", version = "0.14.27", default-features = false, features = ["runtime"] }
 native-tls = { version = "0.2.11", default-features = false, features = ["vendored"] }
 pathdiff = { version = "0.2.1", default-features = false, features = ["camino"] }
 ring = { version = "0.17.7", features = ["std"] }
@@ -298,7 +302,7 @@ security-framework = { version = "2.7.0", features = ["alpn"] }
 [target.x86_64-apple-darwin.build-dependencies]
 bitflags = { version = "2.4.1", default-features = false, features = ["std"] }
 camino = { version = "1.1.7", default-features = false, features = ["serde1"] }
-hyper = { version = "0.14.27", default-features = false, features = ["runtime"] }
+hyper-582f2526e08bb6a0 = { package = "hyper", version = "0.14.27", default-features = false, features = ["runtime"] }
 native-tls = { version = "0.2.11", default-features = false, features = ["vendored"] }
 pathdiff = { version = "0.2.1", default-features = false, features = ["camino"] }
 ring = { version = "0.17.7", features = ["std"] }
